Device data repository. Unit tests

+++ model/device_data.go
@@ -1,8 +1,8 @@
 package model
 
 type DeviceData struct {
-	Id int
-	DeviceId int
-	Timestamp int64
-	Temperature int
+	Id          int
+	DeviceId    int
+	Timestamp   int64
+	Temperature float64
 }
